Adapt to Diff / Conditionals
