+++ package.json
@@ -251,36 +251,36 @@
                     "dace.sdfv.taskletLOD": {
                         "type": "number",
                         "default": 0.35,
-                        "minimum": 0.0,
-                        "maximum": 5.0,
+                        "minimum": 0,
+                        "maximum": 5,
                         "description": "Zoom-level threshold for drawing tasklet contents"
                     },
                     "dace.sdfv.scopeLOD": {
                         "type": "number",
                         "default": 0.75,
-                        "minimum": 0.0,
-                        "maximum": 5.0,
+                        "minimum": 0,
+                        "maximum": 5,
                         "description": "Zoom threshold for drawing simple versions of scopes"
                     },
                     "dace.sdfv.edgeLOD": {
                         "type": "number",
-                        "default": 5.0,
-                        "minimum": 0.0,
-                        "maximum": 10.0,
+                        "default": 5,
+                        "minimum": 0,
+                        "maximum": 10,
                         "description": "Zoom threshold for drawing memlets and control flow edges"
                     },
                     "dace.sdfv.nodeLOD": {
                         "type": "number",
-                        "default": 5.0,
-                        "minimum": 0.0,
-                        "maximum": 10.0,
+                        "default": 5,
+                        "minimum": 0,
+                        "maximum": 10,
                         "description": "Zoom threshold for drawing node shapes and labels"
                     },
                     "dace.sdfv.textLOD": {
                         "type": "number",
                         "default": 1.5,
-                        "minimum": 0.0,
-                        "maximum": 10.0,
+                        "minimum": 0,
+                        "maximum": 10,
                         "description": "Zoom threshold for drawing node text"
                     },
                     "dace.sdfv.nestedLOD": {
@@ -753,45 +753,46 @@
         "postinstall": "patch-package --patch-dir ./packages/sdfv/patches"
     },
     "devDependencies": {
-        "@types/dagre": "^0.7.48",
+        "@types/dagre": "^0.7.52",
         "@types/glob": "^8.0.0",
         "@types/jquery": "^3.5.16",
         "@types/jquery-editable-select": "^2.2.1",
         "@types/mocha": "^10.0.1",
         "@types/node": "^12.20.55",
+        "@types/semver": "^7.5.8",
         "@types/uuid": "^8.3.4",
         "@types/vscode": "^1.75.0",
-        "@typescript-eslint/eslint-plugin": "^6.15.0",
-        "@typescript-eslint/parser": "^6.15.0",
+        "@typescript-eslint/eslint-plugin": "^8.4.0",
+        "@typescript-eslint/parser": "^8.4.0",
         "@vscode/test-electron": "^2.2.2",
         "autoprefixer": "^10.4.13",
         "copy-webpack-plugin": "^11.0.0",
-        "css-loader": "^6.8.1",
-        "eslint": "^8.31.0",
+        "css-loader": "^7.1.2",
+        "eslint": "^9.9.1",
         "file-loader": "^6.2.0",
         "glob": "^8.0.3",
         "mocha": "^10.2.0",
-        "postcss-loader": "^7.3.3",
-        "rimraf": "^5.0.5",
-        "sass": "^1.57.1",
-        "sass-loader": "^13.3.2",
-        "style-loader": "^3.3.3",
+        "postcss-loader": "^8.1.1",
+        "rimraf": "^6.0.1",
+        "sass": "^1.77.6",
+        "sass-loader": "^16.0.1",
+        "style-loader": "^4.0.0",
         "ts-loader": "^9.5.1",
-        "typescript": "^5.3.3",
-        "webpack": "^5.89.0",
+        "typescript": "^5.5.4",
+        "webpack": "^5.94.0",
         "webpack-cli": "^5.1.4"
     },
     "dependencies": {
         "@popperjs/core": "^2.11.6",
         "@vscode/debugadapter": "^1.58.0",
         "@vscode/debugprotocol": "^1.58.0",
-        "bootstrap": "^5.3.0",
+        "bootstrap": "^5.3.3",
         "browserify-zlib": "^0.2.0",
         "dagre": "^0.8.5",
-        "jquery": "^3.6.3",
+        "jquery": "^3.7.1",
         "jquery-editable-select": "^2.2.5",
-        "material-symbols": "^0.19.0",
-        "monaco-editor": "^0.45.0",
+        "material-symbols": "^0.22.2",
+        "monaco-editor": "^0.51.0",
         "process": "^0.11.10",
         "semver": "^7.6.2",
         "stream-browserify": "^3.0.0",
