+++ src/webclients/components/sdfv/vscode_sdfv.ts
@@ -17,29 +17,19 @@ import './vscode_sdfv.scss';
 import * as staticSdfgMetaDict from '../../../utils/sdfg_meta_dict.json';
 
 import {
-    AccessNode,
     DagreGraph,
-    Edge,
-    EntryNode,
-    ExitNode,
-    findGraphElementByUUID,
-    find_in_graph,
-    GenericSdfgOverlay,
     getGraphElementUUID,
     JsonSDFG,
     LogicalGroup,
     LogicalGroupOverlay,
     MemoryLocationOverlay,
     MemoryVolumeOverlay,
-    mouse_event,
-    NestedSDFG,
     OperationalIntensityOverlay,
     SimulatedOperationalIntensityOverlay,
     parse_sdfg,
     Point2D,
     read_or_decompress,
     RuntimeMicroSecondsOverlay,
-    ScopeNode,
     SDFG,
     SDFGElement,
     SDFGRenderer,
@@ -52,7 +42,9 @@ import {
     traverseSDFGScopes,
     JsonSDFGState,
     checkCompatLoad,
-    find_in_graph_predicate,
+    ISDFVUserInterface,
+    findInGraphPredicate,
+    findInGraph,
 } from '@spcl/sdfv/src';
 import { LViewRenderer } from '@spcl/sdfv/src/local_view/lview_renderer';
 import {
@@ -83,11 +75,6 @@ import {
     showTransformationDetails
 } from './transformation/transformation';
 import {
-    appendDataDescriptorTable,
-    appendSymbolsTable,
-    generateAttributesTable
-} from './utils/attributes_table';
-import {
     findJsonSDFGElementByUUID,
     highlightUUIDs,
     isCollapsible,
@@ -100,9 +87,12 @@ import {
 } from './utils/helpers';
 import { ComponentTarget } from '../../../components/components';
 import { gzipSync } from 'zlib';
+import { SDFVVSCodeUI } from './vscode_sdfv_ui';
+import {
+    GenericSdfgOverlay,
+} from '@spcl/sdfv/src/overlays/generic_sdfg_overlay';
 
 declare const vscode: any;
-declare let SPLIT_DIRECTION: 'vertical' | 'horizontal';
 
 export class VSCodeSDFV extends SDFV {
 
@@ -117,6 +107,8 @@ export class VSCodeSDFV extends SDFV {
     public static getInstance(): VSCodeSDFV {
         return this.INSTANCE;
     }
+
+    private readonly UI: SDFVVSCodeUI = SDFVVSCodeUI.getInstance();
 
     public static readonly OVERLAYS: {
         [key: string]: typeof GenericSdfgOverlay,
@@ -135,16 +127,6 @@ export class VSCodeSDFV extends SDFV {
 
     private processingOverlay?: JQuery<HTMLElement>;
     private processingOverlayMsg?: JQuery<HTMLElement>;
-    private infoContainer?: JQuery<HTMLElement>;
-    private layoutToggleBtn?: JQuery<HTMLElement>;
-    private expandInfoBtn?: JQuery<HTMLElement>;
-    private infoCloseBtn?: JQuery<HTMLElement>;
-    private topBar?: JQuery<HTMLElement>;
-
-    private infoDragBar?: JQuery<HTMLElement>;
-    private draggingInfoBar: boolean = false;
-    private infoBarLastVertWidth: number = 350;
-    private infoBarLastHorHeight: number = 200;
 
     private monaco: any | null = null;
     private origSDFG: JsonSDFG | null = null;
@@ -163,11 +145,9 @@ export class VSCodeSDFV extends SDFV {
     };
     private selectedTransformation: JsonTransformation | null = null;
 
-    public infoTrayExplicitlyHidden: boolean = false;
-
     public initialize(): void {
         this.initDOM();
-        this.initInfoBox();
+        this.UI.init();
         this.initSearch();
 
         this.refreshSdfg().then(() => {
@@ -180,156 +160,6 @@ export class VSCodeSDFV extends SDFV {
         this.processingOverlay = $('#processing-overlay');
         this.processingOverlayMsg = $('#processing-overlay-msg');
 
-        this.infoContainer = $('#info-container');
-        this.layoutToggleBtn = $('#layout-toggle-btn');
-        this.infoDragBar = $('#info-drag-bar');
-        this.expandInfoBtn = $('#expand-info-btn');
-        this.infoCloseBtn = $('#info-close-btn');
-        this.topBar = $('#top-bar');
-    }
-
-    private initInfoBox(): void {
-        // Set up resizing of the info drawer.
-        this.draggingInfoBar = false;
-        const infoChangeHeightHandler = (e: any) => {
-            if (this.draggingInfoBar) {
-                const documentHeight = $('body').innerHeight();
-                if (documentHeight) {
-                    const newHeight = documentHeight - e.originalEvent.y;
-                    if (newHeight < documentHeight) {
-                        this.infoBarLastHorHeight = newHeight;
-                        this.infoContainer?.height(
-                            this.infoBarLastHorHeight.toString() + 'px'
-                        );
-                    }
-                }
-            }
-        };
-        const infoChangeWidthHandler = (e: any) => {
-            if (this.draggingInfoBar) {
-                const documentWidth = $('body').innerWidth();
-                if (documentWidth) {
-                    const newWidth = documentWidth - e.originalEvent.x;
-                    if (newWidth < documentWidth) {
-                        this.infoBarLastVertWidth = newWidth;
-                        this.infoContainer?.width(newWidth.toString() + 'px');
-
-                        if (SDFVSettings.get<boolean>('minimap')) {
-                            $('#minimap').css('transition', '');
-                            $('#minimap').css(
-                                'right', (newWidth + 5).toString() + 'px'
-                            );
-                        }
-                    }
-                }
-            }
-        };
-        $(document).on('mouseup', () => {
-            this.draggingInfoBar = false;
-        });
-        this.infoDragBar?.on('mousedown', () => {
-            this.draggingInfoBar = true;
-        });
-        if (SPLIT_DIRECTION === 'vertical')
-            $(document).on('mousemove', infoChangeWidthHandler);
-        else
-            $(document).on('mousemove', infoChangeHeightHandler);
-
-        // Set up changing the info drawer layout.
-        this.layoutToggleBtn?.on('click', () => {
-            const oldDir = SPLIT_DIRECTION;
-            SPLIT_DIRECTION = SPLIT_DIRECTION === 'vertical' ?
-                'horizontal' : 'vertical';
-            this.layoutToggleBtn?.removeClass(oldDir);
-            this.layoutToggleBtn?.addClass(SPLIT_DIRECTION);
-            if (oldDir === 'vertical') {
-                this.infoContainer?.removeClass('offcanvas-end');
-                this.infoContainer?.addClass('offcanvas-bottom');
-                this.infoDragBar?.removeClass('gutter-vertical');
-                this.infoDragBar?.addClass('gutter-horizontal');
-                this.expandInfoBtn?.html(
-                    '<span><i class="material-symbols-outlined">' +
-                        'bottom_panel_open</i></span>'
-                );
-                $(document).off('mousemove', infoChangeWidthHandler);
-                $(document).on('mousemove', infoChangeHeightHandler);
-                this.infoContainer?.width('100%');
-                this.infoContainer?.height(
-                    this.infoBarLastHorHeight.toString() + 'px'
-                );
-            } else {
-                this.infoContainer?.removeClass('offcanvas-bottom');
-                this.infoContainer?.addClass('offcanvas-end');
-                this.infoDragBar?.removeClass('gutter-horizontal');
-                this.infoDragBar?.addClass('gutter-vertical');
-                this.expandInfoBtn?.html(
-                    '<span><i class="material-symbols-outlined">' +
-                        'right_panel_open</i></span>'
-                );
-                $(document).off('mousemove', infoChangeHeightHandler);
-                $(document).on('mousemove', infoChangeWidthHandler);
-                this.infoContainer?.height('100%');
-                this.infoContainer?.width(
-                    this.infoBarLastVertWidth.toString() + 'px'
-                );
-            }
-
-            infoBoxCheckStacking(this.infoContainer);
-            infoBoxCheckUncoverTopBar(this.infoContainer, this.topBar);
-            this.checkTrayCoversMinimap();
-
-            SDFVComponent.getInstance().invoke(
-                'setSplitDirection', [SPLIT_DIRECTION]
-            );
-        });
-
-        if (this.infoContainer)
-            new ResizeObserver(() => {
-                infoBoxCheckStacking(this.infoContainer);
-            }).observe(this.infoContainer[0]);
-
-        // Set up toggling the info tray.
-        this.infoCloseBtn?.on('click', () => {
-            this.expandInfoBtn?.show();
-            this.infoContainer?.removeClass('show');
-            VSCodeSDFV.getInstance().infoTrayExplicitlyHidden = true;
-            this.checkTrayCoversMinimap(true);
-        });
-        this.expandInfoBtn?.on('click', () => {
-            this.expandInfoBtn?.hide();
-            infoBoxCheckUncoverTopBar(this.infoContainer, this.topBar);
-            this.infoContainer?.addClass('show');
-            VSCodeSDFV.getInstance().infoTrayExplicitlyHidden = false;
-            this.checkTrayCoversMinimap(true);
-        });
-    }
-
-    private checkTrayCoversMinimap(animate: boolean = false): void {
-        if (SDFVSettings.get<boolean>('minimap')) {
-            if (SPLIT_DIRECTION === 'vertical' && this.infoBarLastVertWidth &&
-                !this.infoTrayExplicitlyHidden) {
-                try {
-                    const pixels = this.infoBarLastVertWidth + 5;
-                    if (animate)
-                        $('#minimap').css(
-                            'transition', 'right 0.3s ease-in-out'
-                        );
-                    else
-                        $('#minimap').css('transition', '');
-                    $('#minimap').css('right', pixels.toString() + 'px');
-                } catch (e) {
-                    console.warn(e);
-                }
-            } else {
-                if (animate)
-                    $('#minimap').css(
-                        'transition', 'right 0.3s ease-in-out'
-                    );
-                else
-                    $('#minimap').css('transition', '');
-                $('#minimap').css('right', '5px');
-            }
-        }
     }
 
     public initSearch(): void {
@@ -378,9 +208,7 @@ export class VSCodeSDFV extends SDFV {
                     const code = $('#advsearch').val();
                     if (graph && code) {
                         const predicate = eval(code.toString());
-                        find_in_graph_predicate(
-                            this, renderer, graph, predicate
-                        );
+                        findInGraphPredicate(this.UI, renderer, predicate);
                     }
                 }, 1);
             }
@@ -388,82 +216,8 @@ export class VSCodeSDFV extends SDFV {
         });
     }
 
-    public init_menu(): void {
-        this.initInfoBox();
-    }
-
-    public close_menu(): void {
-        this.clearInfoBox();
-    }
-
-    public sidebar_get_contents(): HTMLElement | null {
-        return this.infoBoxGetContents();
-    }
-
-    public sidebar_show(): void {
-        this.infoBoxShow();
-    }
-
-    public sidebar_set_title(title: string): void {
-        this.infoBoxSetTitle(title);
-    }
-
-    public fill_info(elem: SDFGElement): void {
-        this.fillInfo(elem);
-    }
-
     public start_find_in_graph(): void {
         this.startFindInGraph();
-    }
-
-    /**
-     * Get the current info-box contents.
-     */
-    public infoBoxGetContents(): HTMLElement | null {
-        return document.getElementById('info-contents');
-    }
-
-    /**
-     * Show the info box and its necessary components.
-     */
-    public infoBoxShow(overrideHidden: boolean = false): void {
-        if (!this.infoTrayExplicitlyHidden || overrideHidden) {
-            const infoBox = $('#info-container');
-            infoBoxCheckUncoverTopBar(infoBox, $('#top-bar'));
-            infoBox.addClass('show');
-            this.checkTrayCoversMinimap(true);
-        }
-
-        if (SPLIT_DIRECTION === 'vertical')
-            this.infoContainer?.width(
-                this.infoBarLastVertWidth.toString() + 'px'
-            );
-        else
-            this.infoContainer?.height(
-                this.infoBarLastHorHeight.toString() + 'px'
-            );
-    }
-
-    /**
-     * Set the header/title of the info-box in the embedded view.
-     */
-    public infoBoxSetTitle(title: string): void {
-        $('#info-title').text(title);
-    }
-
-    /**
-     * Clear the info container and its title, and hide the clear button again.
-     */
-    public clearInfoBox(hide: boolean = false): void {
-        $('#info-contents').html('');
-        $('#info-title').text('');
-        $('#goto-source-btn').hide();
-        $('#goto-cpp-btn').hide();
-        $('#goto-edge-start').hide();
-        $('#goto-edge-end').hide();
-        this.selectedTransformation = null;
-        if (hide)
-            $('#info-container').removeClass('show');
     }
 
     private handleShowGroupsContextMenu(
@@ -581,32 +335,29 @@ export class VSCodeSDFV extends SDFV {
         sdfv: SDFV,
         endsPan: boolean
     ): boolean {
-        const externalRet = mouse_event(
-            evtype, event, mousepos, elements, renderer, selectedElements, sdfv,
-            endsPan
-        );
-
         if (evtype === 'click' || evtype === 'dblclick') {
-            sdfv.sidebar_show();
+            this.UI.infoShow();
         } else if (evtype === 'contextmenu') {
-            if (VSCodeRenderer.getInstance()?.get_overlay_manager()
-                .is_overlay_active(LogicalGroupOverlay))
+            if (VSCodeRenderer.getInstance()?.overlayManager.is_overlay_active(
+                LogicalGroupOverlay
+            )) {
                 VSCodeSDFV.getInstance().handleShowGroupsContextMenu(
                     selectedElements, event
                 );
+            }
 
             event.preventDefault();
             return false;
         }
 
-        return externalRet;
+        return false;
     }
 
     @ICPCRequest()
     public async getCompressedSDFG(): Promise<Uint8Array | null> {
         const sdfgString = this.getSdfgString();
         if (sdfgString)
-            return gzipSync(sdfgString);
+            return new Uint8Array(gzipSync(sdfgString));
         return null;
     }
 
@@ -713,150 +464,6 @@ export class VSCodeSDFV extends SDFV {
         );
     }
 
-    /**
-     * Fill out the info-box of the embedded layout with info about an element.
-     * This dynamically builds one or more tables showing all of the relevant
-     * info about a given element.
-     */
-    public fillInfo(elem: SDFGElement): void {
-        const buttons = [
-            $('#goto-source-btn'),
-            $('#goto-cpp-btn'),
-            $('#goto-edge-start'),
-            $('#goto-edge-end'),
-        ];
-
-        // Clear and hide these buttons.
-        buttons.forEach((btn) => {
-            btn.hide();
-            btn.off('click');
-            btn.prop('title', '');
-        });
-
-        if (elem) {
-            this.infoBoxSetTitle(elem.type() + ' ' + elem.label());
-
-            const contents = $('#info-contents');
-            contents.html('');
-
-            if (elem instanceof Edge && elem.data.type === 'Memlet' &&
-                elem.parent_id !== null) {
-                const ndEdges = elem.cfg?.nodes[elem.parent_id].edges;
-                if (ndEdges) {
-                    let sdfg_edge = ndEdges[elem.id];
-                    $('<p>', {
-                        'class': 'info-subtitle',
-                        'html': 'Connectors: ' + sdfg_edge.src_connector +
-                            ' <i class="material-symbols-outlined">' +
-                            'arrow_forward</i> ' + sdfg_edge.dst_connector,
-                    }).appendTo(contents);
-                    $('<hr>').appendTo(contents);
-                }
-            }
-
-            const tableContainer = $('<div>', {
-                'class': 'container-fluid attr-table-base-container',
-            }).appendTo(contents);
-            generateAttributesTable(elem, undefined, tableContainer);
-
-            if (elem instanceof AccessNode) {
-                // If we're processing an access node, add array info too.
-                const sdfg_array = elem.sdfg.attributes._arrays[
-                    elem.attributes().data
-                ];
-                $('<br>').appendTo(contents);
-                $('<p>', {
-                    'class': 'info-subtitle',
-                    'text': sdfg_array.type + ' properties:',
-                }).appendTo(contents);
-
-                // TODO: Allow container types to be changed here too.
-                const tableContainer = $('<div>', {
-                    'class': 'container-fluid attr-table-base-container',
-                }).appendTo(contents);
-                generateAttributesTable(sdfg_array, undefined, tableContainer);
-            } else if (elem instanceof NestedSDFG) {
-                // If nested SDFG, add SDFG info too.
-                const sdfg_sdfg = elem.attributes().sdfg;
-                $('<br>').appendTo(contents);
-                $('<p>', {
-                    'class': 'info-subtitle',
-                    'text': 'SDFG properties:',
-                }).appendTo(contents);
-
-                const tableContainer = $('<div>', {
-                    'class': 'container-fluid attr-table-base-container',
-                }).appendTo(contents);
-                generateAttributesTable(sdfg_sdfg, undefined, tableContainer);
-            } else if (elem instanceof ScopeNode) {
-                // If we're processing a scope node, we want to append the exit
-                // node's props when selecting an entry node, and vice versa.
-                let other_element = undefined;
-
-                let other_uuid = undefined;
-                if (elem instanceof EntryNode &&
-                    !elem.attributes().is_collapsed) {
-                    other_uuid = elem.cfg!.cfg_list_id + '/' +
-                        elem.parent_id + '/' +
-                        elem.data.node.scope_exit + '/-1';
-                } else if (elem instanceof ExitNode) {
-                    other_uuid = elem.cfg!.cfg_list_id + '/' +
-                        elem.parent_id + '/' +
-                        elem.data.node.scope_entry + '/-1';
-                }
-
-                if (other_uuid) {
-                    other_element = findGraphElementByUUID(
-                        VSCodeRenderer.getInstance()!.getCFGList(),
-                        other_uuid
-                    );
-                }
-
-                if (other_element && other_element instanceof SDFGElement) {
-                    $('<br>').appendTo(contents);
-                    $('<p>', {
-                        'class': 'info-subtitle',
-                        'text':
-                            other_element.type() + ' ' + other_element.label(),
-                    }).appendTo(contents);
-
-                    const tableContainer = $('<div>', {
-                        'class': 'container-fluid attr-table-base-container',
-                    }).appendTo(contents);
-                    generateAttributesTable(
-                        other_element, undefined, tableContainer
-                    );
-                }
-            } else if (elem instanceof SDFG) {
-                if (elem.data?.attributes) {
-                    appendDataDescriptorTable(
-                        contents, elem.data.attributes._arrays, elem.data
-                    );
-                    appendSymbolsTable(
-                        contents, elem.data.attributes.symbols, elem.data
-                    );
-                }
-            } else if (elem instanceof NestedSDFG) {
-                if (elem.data?.node?.attributes) {
-                    appendDataDescriptorTable(
-                        contents,
-                        elem.data.node.attributes.sdfg.attributes._arrays,
-                        elem.data.node.attributes.sdfg
-                    );
-                    appendSymbolsTable(
-                        contents,
-                        elem.data.node.attributes.sdfg.attributes.symbols,
-                        elem.data.node.attributes.sdfg
-                    );
-                }
-            }
-
-            infoBoxCheckStacking($('#info-container'));
-        } else {
-            this.clearInfoBox();
-        }
-    }
-
     public startFindInGraph(): void {
         const renderer = VSCodeRenderer.getInstance();
         if (renderer)
@@ -865,8 +472,8 @@ export class VSCodeSDFV extends SDFV {
                 const graph = renderer.get_graph();
                 if (graph && searchVal !== undefined &&
                     typeof searchVal === 'string' && searchVal.length > 0)
-                    find_in_graph(
-                        this, renderer, graph, searchVal,
+                    findInGraph(
+                        this.UI, renderer, searchVal,
                         $('#search-case').is(':checked')
                     );
             }, 1);
@@ -924,8 +531,8 @@ export class VSCodeSDFV extends SDFV {
         if (selectedElements && selectedElements.length === 1)
             reselectRendererElement(selectedElements[0]);
         else if (!selectedElements || selectedElements.length === 0)
-            this.fillInfo(
-                new SDFG(this.renderer.get_sdfg())
+            this.UI.showElementInfo(
+                new SDFG(this.renderer.get_sdfg()), this.renderer
             );
     }
 
@@ -1095,7 +702,7 @@ export class VSCodeSDFV extends SDFV {
     public setShowingBreakpoints(showingBreakpoints: boolean): void {
         this.showingBreakpoints = showingBreakpoints;
         const alreadyActive =
-            VSCodeRenderer.getInstance()?.get_overlay_manager().get_overlay(
+            VSCodeRenderer.getInstance()?.overlayManager.get_overlay(
                 BreakpointIndicator
             );
         if (this.showingBreakpoints && alreadyActive === undefined) {
@@ -1141,7 +748,7 @@ export class VSCodeSDFV extends SDFV {
             this.renderer = null;
         }
         this.localViewRenderer = localViewRenderer;
-        this.infoBoxShow(true);
+        this.UI.infoShow(true);
     }
 
     @ICPCRequest()
@@ -1173,7 +780,7 @@ export class VSCodeSDFV extends SDFV {
             this.setRendererContent(pSdfg, true);
             $('#exit-preview-button')?.show();
             if (histIndex !== undefined) {
-                this.clearInfoBox();
+                this.UI.infoClear();
                 this.setViewingHistoryState(true, histIndex);
                 refreshTransformationList();
             }
@@ -1254,9 +861,7 @@ export class VSCodeSDFV extends SDFV {
 
     @ICPCRequest()
     public async setOverlays(overlays: string[]): Promise<void> {
-        const overlayManager =
-            VSCodeRenderer.getInstance()?.get_overlay_manager();
-        const activeOverlays = overlayManager?.get_overlays();
+        const overlayManager = VSCodeRenderer.getInstance()?.overlayManager;
 
         const toActivate = [];
         for (const ol of overlays)
@@ -1281,6 +886,10 @@ export class VSCodeSDFV extends SDFV {
         );
         this.setRendererContent(specialized, false, false);
         this.setProcessingOverlay(false);
+    }
+
+    public get linkedUI(): ISDFVUserInterface {
+        return this.UI;
     }
 
 }
@@ -1319,32 +928,6 @@ export class SDFVComponent extends ICPCWebclientMessagingComponent {
         );
     }
 
-}
-
-function infoBoxCheckUncoverTopBar(
-    infoContainer?: JQuery<HTMLElement>, topBar?: JQuery<HTMLElement>
-): void {
-    // If the info container is to the side, ensure it doesn't cover up the
-    // top bar when shown.
-    if (infoContainer?.hasClass('offcanvas-end')) {
-        const topBarHeight = topBar?.outerHeight(false);
-        infoContainer?.css('top', topBarHeight + 'px');
-    } else {
-        infoContainer?.css('top', '');
-    }
-}
-
-/**
- * Check if the info box is wide enough to show keys / values side-by-side.
- * If not, stack them one on top of the other.
- * @param infoContainer The info box container.
- */
-function infoBoxCheckStacking(infoContainer?: JQuery<HTMLElement>): void {
-    const innerWidth = infoContainer?.innerWidth();
-    if (innerWidth && innerWidth <= 575)
-        infoContainer?.addClass('stacked');
-    else
-        infoContainer?.removeClass('stacked');
 }
 
 $(() => {
