+++ tsconfig.json
@@ -1,10 +1,10 @@
 {
 	"compilerOptions": {
 		"module": "commonjs",
-		"target": "ES2022",
+		"target": "ESNext",
 		"outDir": "out",
 		"lib": [
-			"ES2022",
+			"ESNext",
 			"dom"
 		],
 		"sourceMap": true,
